+++ packages/@kagal-ed25519-secret/src/index.ts
@@ -4,6 +4,16 @@ import pkg from '../package.json' with { type: 'json' };
 export const VERSION: string = pkg.version;
 
 export {
+  asEd25519Seed,
+  type Ed25519Seed,
+  type KeyPair,
+  newKeyPair,
+} from './key';
+export {
+  type KeyConfig,
+  parseSecretToKey,
+} from './secret';
+export {
   assertValidSelector,
   isValidSelector,
   SELECTOR_PATTERN,
@@ -12,3 +22,9 @@ export {
   newSigner,
   type Signer,
 } from './signer';
+export {
+  asBytes,
+  decodeBase64,
+  encodeBase64,
+  getRandom,
+} from './utils';